Fix investigation UI lint issues
